Backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
gunicorn
transformers
langchain-community
sentence-transformers
chromadb
huggingface-hub
pypdf